Deprecate the misnamed YOTTA_CFG_MTRACE_TMP_LINE_LENGTH flag, which is to be removed in the next release.
Configuring the tmp line length value through config.json in yotta requires defining a separate "mtrace" group,
because the name of the flag doesn't follow the 'YOTTA_CFG_MBED_TRACE_' pattern.

This allows configuring the tmp line length in the same group mbed-trace is enabled in.

(The name of the flag seems to be a relic of the time the name of the library was still being decided.
For some reason it wasn't changed with the other ones when 'mbed-trace' was decided on.)