Fix disabling traces through the YOTTA_CFG_MBED_TRACE flag.
Previously only the existence of the flag was checked. Thus tracing was always enabled if the flag was defined
regardless of its actual value, including when it was defined as 0 or false.

Now tracing is only enabled when YOTTA_CFG_MBED_TRACE is defined as 1 or true.
Tracing is disabled when the flag is undefined, or defined as 0 or false.