Fix support for TRACE_CARRIAGE_RETURN
The option TRACE_CARRIAGE_RETURN is enabled in the default configuration
of mbed-trace.

It is an ambiguous option. In most of the cases (see mbed_vtracef()) and
despite its name, it adds a new line (\n) at the end of each trace.

However, there are two bugs that cancel each other:
  - by default, mbed-trace use puts() as backend. puts() already add a
    new line at end of the string displayed. So this behavior is not
    coherent with DEFAULT_TRACE_CONFIG.
  - by default, TRACE_MODE_COLOR is also enabled. In this case (and only
    in this case), TRACE_CARRIAGE_RETURN appends a \r instead of a \n.

This patch fixes the case where TRACE_MODE_COLOR and
TRACE_CARRIAGE_RETURN are both enabled. We also remove the original ANSI
escape sequence that make no sense.

Then, to be coherent with the value of DEFAULT_TRACE_CONFIG, the patch
replaces puts() by printf(). It could have change the value of
DEFAULT_TRACE_CONFIG but I think this way has less probability to break
applications.

v4:
  - fix printf usage (note: it is also possible to use fputs(str, stdout),
    any preference?)
v3:
  - fix previous rebase
v2:
  - rebase upon master

Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>